Split logic into various files for maintainability
