+++ internal/controller/module_controller.go
@@ -17,13 +17,9 @@ limitations under the License.
 package controller
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net/http"
 	"time"
 
 	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
@@ -37,7 +33,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/yaml"
 
 	batchv1 "github.com/forkspacer/forkspacer/api/v1"
 	"github.com/forkspacer/forkspacer/pkg/manager"
@@ -131,45 +126,6 @@ func (r *ModuleReconciler) SetupWithManager(mgr ctrl.Manager) error {
 			MaxConcurrentReconciles: 10,
 		}).
 		Complete(r)
-}
-
-func (r *ModuleReconciler) handleDeletion(ctx context.Context, module *batchv1.Module) (ctrl.Result, error) {
-	log := logf.FromContext(ctx)
-
-	switch module.Status.Phase {
-	case batchv1.ModulePhaseReady, batchv1.ModulePhaseFailed, batchv1.ModulePhaseSleeped:
-	default:
-		return ctrl.Result{RequeueAfter: time.Second * 2}, nil
-	}
-
-	if err := r.setPhase(ctx, module, batchv1.ModulePhaseUninstalling, nil); err != nil {
-		return ctrl.Result{}, err
-	}
-
-	modulePhase := module.Status.Phase
-	if err := r.uninstallModule(ctx, module); err != nil {
-		log.Error(err, "module uninstallation failed")
-		if modulePhase != batchv1.ModulePhaseFailed {
-			r.Recorder.Event(module, "Warning", "UninstallError", fmt.Sprintf("Module uninstallation failed: %v", err))
-			return ctrl.Result{}, err
-		}
-	}
-
-	return ctrl.Result{}, retry.RetryOnConflict(
-		retry.DefaultRetry,
-		func() error {
-			if err := r.Get(ctx, client.ObjectKeyFromObject(module), module); err != nil {
-				return err
-			}
-
-			if !controllerutil.ContainsFinalizer(module, moduleFinalizers.Delete) {
-				return nil
-			}
-
-			controllerutil.RemoveFinalizer(module, moduleFinalizers.Delete)
-			return r.Update(ctx, module)
-		},
-	)
 }
 
 func (r *ModuleReconciler) handleEmptyPhase(ctx context.Context, module *batchv1.Module) (ctrl.Result, error) {
@@ -275,259 +231,6 @@ func (r *ModuleReconciler) setPhase(ctx context.Context, module *batchv1.Module,
 	)
 }
 
-func (r *ModuleReconciler) installModule(ctx context.Context, module *batchv1.Module) error {
-	log := logf.FromContext(ctx)
-
-	workspace, err := r.getWorkspace(ctx, &module.Spec.Workspace)
-	if err != nil {
-		return fmt.Errorf("failed to get workspace: %v", err)
-	}
-
-	if !workspace.Status.Ready {
-		return fmt.Errorf("workspace not ready: workspace %s/%s is not ready", workspace.Namespace, workspace.Name)
-	}
-
-	if workspace.Status.Phase != batchv1.WorkspacePhaseReady {
-		return fmt.Errorf(
-			"workspace not in running phase: workspace %s/%s is not in '%s' phase, current phase is %s",
-			workspace.Namespace, workspace.Name, batchv1.WorkspacePhaseReady, workspace.Status.Phase,
-		)
-	}
-
-	moduleReader, err := r.readModuleLocation(module.Spec.Source)
-	if err != nil {
-		return fmt.Errorf("failed to read module location: %v", err)
-	}
-
-	moduleData, err := io.ReadAll(moduleReader)
-	if err != nil {
-		return fmt.Errorf("failed to read module data: %v", err)
-	}
-
-	annotations := map[string]string{
-		types.ModuleAnnotationKeys.Resource: string(moduleData),
-	}
-
-	metaData := make(managerBase.MetaData)
-
-	configMap := make(map[string]any)
-	if module.Spec.Config != nil && module.Spec.Config.Raw != nil {
-		if err := json.Unmarshal(module.Spec.Config.Raw, &configMap); err != nil {
-			return fmt.Errorf("failed to unmarshal module config: %v", err)
-		}
-
-		annotations[types.ModuleAnnotationKeys.BaseModuleConfig] = string(module.Spec.Config.Raw)
-	}
-
-	if iManager, err := r.newManager(ctx, module, workspace.Spec.Connection, moduleData, metaData, configMap); err != nil {
-		return err
-	} else {
-		installErr := iManager.Install(ctx, metaData)
-
-		patch := client.MergeFrom(module.DeepCopy())
-		annotations[types.ModuleAnnotationKeys.ManagerData] = metaData.String()
-		utils.UpdateMap(&module.Annotations, annotations)
-		if err = r.Patch(ctx, module, patch); err != nil {
-			log.Error(err, "failed to patch module with install annotations", "module", module.Name, "namespace", module.Namespace)
-		}
-
-		if installErr != nil {
-			return fmt.Errorf("failed to install module: %w", installErr)
-		}
-		return nil
-	}
-}
-
-func (r *ModuleReconciler) uninstallModule(ctx context.Context, module *batchv1.Module) error {
-	log := logf.FromContext(ctx)
-
-	workspace, err := r.getWorkspace(ctx, &module.Spec.Workspace)
-	if err != nil {
-		return fmt.Errorf("failed to get workspace for module %s/%s: %v", module.Namespace, module.Name, err)
-	}
-
-	if !workspace.Status.Ready {
-		return fmt.Errorf("workspace not ready: workspace %s/%s is not ready", workspace.Namespace, workspace.Name)
-	}
-
-	switch workspace.Status.Phase {
-	case batchv1.WorkspacePhaseReady, batchv1.WorkspacePhaseHibernated, batchv1.WorkspacePhaseFailed:
-	default:
-		return fmt.Errorf(
-			"workspace is not in a suitable phase for uninstallation: expected one of ['%s', '%s', '%s'], but got '%s'",
-			batchv1.WorkspacePhaseReady, batchv1.WorkspacePhaseHibernated, batchv1.WorkspacePhaseFailed, workspace.Status.Phase,
-		)
-	}
-
-	utils.InitMap(&module.Annotations)
-
-	resourceAnnotation := module.Annotations[types.ModuleAnnotationKeys.Resource]
-	if resourceAnnotation == "" {
-		return fmt.Errorf("resource definition not found in module annotations for %s/%s", module.Namespace, module.Name)
-	}
-	moduleData := []byte(resourceAnnotation)
-
-	managerData, ok := module.Annotations[types.ModuleAnnotationKeys.ManagerData]
-
-	metaData := make(managerBase.MetaData)
-	if ok && managerData != "" {
-		err := metaData.Parse([]byte(managerData))
-		if err != nil {
-			log.Error(err, "failed to parse manager data for module", "module", module.Name, "namespace", module.Namespace)
-			// Proceed with uninstall even if metadata parsing fails, as it might not be critical for uninstall
-		}
-	}
-
-	configMap := make(map[string]any)
-	configMapData, ok := module.Annotations[types.ModuleAnnotationKeys.BaseModuleConfig]
-	if ok {
-		if err := json.Unmarshal([]byte(configMapData), &configMap); err != nil {
-			log.Error(err, "failed to unmarshal module config from annotation")
-		}
-	}
-
-	if iManager, err := r.newManager(ctx, module, workspace.Spec.Connection, moduleData, metaData, configMap); err != nil {
-		return err
-	} else {
-		if err := iManager.Uninstall(ctx, metaData); err != nil {
-			return fmt.Errorf("failed to uninstall module: %w", err)
-		}
-		return nil
-	}
-}
-
-func (r *ModuleReconciler) sleepModule(ctx context.Context, module *batchv1.Module) error {
-	log := logf.FromContext(ctx)
-
-	workspace, err := r.getWorkspace(ctx, &module.Spec.Workspace)
-	if err != nil {
-		return fmt.Errorf("failed to get workspace for module %s/%s: %v", module.Namespace, module.Name, err)
-	}
-
-	if !workspace.Status.Ready {
-		return fmt.Errorf("workspace not ready: workspace %s/%s is not ready", workspace.Namespace, workspace.Name)
-	}
-
-	switch workspace.Status.Phase {
-	case batchv1.WorkspacePhaseReady, batchv1.WorkspacePhaseHibernated:
-	default:
-		return fmt.Errorf(
-			"workspace is not in a suitable phase for sleeping: expected one of ['%s', '%s'], but got '%s'",
-			batchv1.WorkspacePhaseReady, batchv1.WorkspacePhaseHibernated, workspace.Status.Phase,
-		)
-	}
-
-	utils.InitMap(&module.Annotations)
-
-	resourceAnnotation := module.Annotations[types.ModuleAnnotationKeys.Resource]
-	if resourceAnnotation == "" {
-		return fmt.Errorf("resource definition not found in module annotations for %s/%s", module.Namespace, module.Name)
-	}
-	moduleData := []byte(resourceAnnotation)
-
-	managerData, ok := module.Annotations[types.ModuleAnnotationKeys.ManagerData]
-
-	metaData := make(managerBase.MetaData)
-	if ok && managerData != "" {
-		err := metaData.Parse([]byte(managerData))
-		if err != nil {
-			log.Error(err, "failed to parse manager data for module", "module", module.Name, "namespace", module.Namespace)
-			// Proceed with sleep even if metadata parsing fails, as it might not be critical for sleep
-		}
-	}
-
-	configMap := make(map[string]any)
-	configMapData, ok := module.Annotations[types.ModuleAnnotationKeys.BaseModuleConfig]
-	if ok {
-		if err := json.Unmarshal([]byte(configMapData), &configMap); err != nil {
-			log.Error(err, "failed to unmarshal module config from annotation")
-		}
-	}
-
-	if iManager, err := r.newManager(ctx, module, workspace.Spec.Connection, moduleData, metaData, configMap); err != nil {
-		return err
-	} else {
-		sleepErr := iManager.Sleep(ctx, metaData)
-
-		patch := client.MergeFrom(module.DeepCopy())
-		module.Annotations[types.ModuleAnnotationKeys.ManagerData] = metaData.String()
-		if err := r.Patch(ctx, module, patch); err != nil {
-			log.Error(err, "failed to patch module with meta data annotations after sleep", "module", module.Name, "namespace", module.Namespace)
-		}
-
-		if sleepErr != nil {
-			return fmt.Errorf("failed to sleep module: %w", sleepErr)
-		}
-		return nil
-	}
-}
-
-func (r *ModuleReconciler) resumeModule(ctx context.Context, module *batchv1.Module) error {
-	log := logf.FromContext(ctx)
-
-	workspace, err := r.getWorkspace(ctx, &module.Spec.Workspace)
-	if err != nil {
-		return fmt.Errorf("failed to get workspace for module %s/%s: %v", module.Namespace, module.Name, err)
-	}
-
-	if !workspace.Status.Ready {
-		return fmt.Errorf("workspace not ready: workspace %s/%s is not ready", workspace.Namespace, workspace.Name)
-	}
-
-	switch workspace.Status.Phase {
-	case batchv1.WorkspacePhaseReady, batchv1.WorkspacePhaseHibernated:
-	default:
-		return fmt.Errorf(
-			"workspace is not in a suitable phase for sleeping: expected one of ['%s', '%s'], but got '%s'",
-			batchv1.WorkspacePhaseReady, batchv1.WorkspacePhaseHibernated, workspace.Status.Phase,
-		)
-	}
-
-	utils.InitMap(&module.Annotations)
-
-	resourceAnnotation := module.Annotations[types.ModuleAnnotationKeys.Resource]
-	if resourceAnnotation == "" {
-		return fmt.Errorf("resource definition not found in module annotations for %s/%s", module.Namespace, module.Name)
-	}
-	moduleData := []byte(resourceAnnotation)
-
-	managerData, ok := module.Annotations[types.ModuleAnnotationKeys.ManagerData]
-
-	metaData := make(managerBase.MetaData)
-	if ok && managerData != "" {
-		err := metaData.Parse([]byte(managerData))
-		if err != nil {
-			log.Error(err, "failed to parse manager data for module", "module", module.Name, "namespace", module.Namespace)
-			// Proceed with resume even if metadata parsing fails, as it might not be critical for resume
-		}
-	}
-
-	configMap := make(map[string]any)
-	configMapData, ok := module.Annotations[types.ModuleAnnotationKeys.BaseModuleConfig]
-	if ok {
-		if err := json.Unmarshal([]byte(configMapData), &configMap); err != nil {
-			log.Error(err, "failed to unmarshal module config from annotation")
-		}
-	}
-
-	if iManager, err := r.newManager(ctx, module, workspace.Spec.Connection, moduleData, metaData, configMap); err != nil {
-		return err
-	} else {
-		resumeErr := iManager.Resume(ctx, metaData)
-
-		patch := client.MergeFrom(module.DeepCopy())
-		module.Annotations[types.ModuleAnnotationKeys.ManagerData] = metaData.String()
-		if err := r.Patch(ctx, module, patch); err != nil {
-			log.Error(err, "failed to patch module with meta data annotations after resume", "module", module.Name, "namespace", module.Namespace)
-		}
-
-		if resumeErr != nil {
-			return fmt.Errorf("failed to resume module: %w", resumeErr)
-		}
-		return nil
-	}
-}
-
 func (r *ModuleReconciler) newManager(
 	ctx context.Context,
 	module *batchv1.Module,
@@ -606,29 +309,4 @@ func (r *ModuleReconciler) getWorkspace(ctx context.Context, workspaceRef *batch
 	}
 
 	return workspace, nil
-}
-
-func (r *ModuleReconciler) readModuleLocation(moduleSource batchv1.ModuleSource) (io.Reader, error) {
-	if moduleSource.Raw != nil {
-		// runtime.RawExtension stores data as JSON, convert it to YAML
-		yamlData, err := yaml.JSONToYAML(moduleSource.Raw.Raw)
-		if err != nil {
-			return nil, fmt.Errorf("failed to convert raw JSON to YAML: %w", err)
-		}
-		return bytes.NewReader(yamlData), nil
-	} else if moduleSource.HttpURL != nil {
-		resp, err := http.Get(*moduleSource.HttpURL)
-		if err != nil {
-			return nil, fmt.Errorf("failed to fetch module from HTTP URL %s: %w", *moduleSource.HttpURL, err)
-		}
-
-		if resp.StatusCode != http.StatusOK {
-			return nil, fmt.Errorf("HTTP request failed with status %d for URL %s", resp.StatusCode, *moduleSource.HttpURL)
-		}
-		return resp.Body, nil
-	} else if moduleSource.Github != nil {
-		return nil, errors.New("'github' module location type is not yet supported")
-	} else {
-		return nil, errors.New("exactly one of 'raw', 'httpURL', or 'github' must be specified")
-	}
 }
